Add files via upload
With this commit, I am uploading my notes and solutions for the first box of Hack The Box called ‘Fawn’. Fawn is a relatively simple machine and is a good starter for learning basic concepts of penetration testing. The box requires minimal enumeration and the exploitation of a weak anonymous FTP service for the initial foothold. The writing is a progress of my personal endeavor on learning and documenting the methodologies of penetration testing. The processes included are discovering active services, ftp access, and file retrieval.